+++ packages/stryker-elements/src/stories/organisms/top-bar.stories.ts
@@ -1,0 +1,10 @@
+import { Meta, StoryObj } from '@storybook/web-components';
+
+import '../../exports/lib/organisms/top-bar';
+
+export default {
+  title: 'Organisms/Top Bar',
+  component: 'sme-top-bar',
+} as Meta;
+
+export const Default: StoryObj = {};