+++ packages/data-access/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.2](https://github.com/stryker-mutator/stryker-dashboard/compare/v0.14.1...v0.14.2) (2024-03-26)
+
+### Bug Fixes
+
+- add healthcheck to docker image ([#536](https://github.com/stryker-mutator/stryker-dashboard/issues/536)) ([813f564](https://github.com/stryker-mutator/stryker-dashboard/commit/813f5646413202b3f2d9233e2ba7096c64fd338e))
+
 ## [0.14.1](https://github.com/stryker-mutator/stryker-dashboard/compare/v0.14.0...v0.14.1) (2024-03-08)
 
 **Note:** Version bump only for package @stryker-mutator/dashboard-data-access